+++ Python Server/app/utils/lang_utils.py
@@ -5,5 +5,7 @@ def get_response_text(status_code):
         return "Added to your mind.", "success"
     elif status_code == 404:
         return "No Notion credentials where provided. Please add them on the server.", "error"
+    elif status_code == 204:
+        return "Title and Url was modified correctly", "success"
     else:
         return "Invalid url or text was provided.", "error"